requirements.txt
django==5.2.10
djangorestframework==3.16.1
psycopg2-binary
